DFLOW-358 Extract copyrighted jobs into CSV file
Create a CSV file containing all copyrighted jobs with their job_id, title and authors.
This is used to check if author is still alive, so that the copyright protection parameter
can be removed.

File is created from file path proviced in ENV variable with current date inserted into
the filename.